requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
python-dateutil>=2.8.0
pyodbc>=4.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
urllib3>=1.26.0
